Revert "ResourceHandle : Refactor resourceHandle data type to long"
Revert submission 28771115-cherrypicker-L41100030005737508:N83900030090946413

Reason for revert: Droidmonitor created revert due to b/360262704. Will be verifying through ABTD before submission.


Reverted changes: /q/submissionid:28771115-cherrypicker-L41100030005737508:N83900030090946413

+++ services/core/java/com/android/server/tv/tunerresourcemanager/LnbResource.java
@@ -37,7 +37,8 @@ public final class LnbResource extends TunerResourceBasic {
      * Builder class for {@link LnbResource}.
      */
     public static class Builder extends TunerResourceBasic.Builder {
-        Builder(long handle) {
+
+        Builder(int handle) {
             super(handle);
         }
